% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIT.R
\name{convert.expression}
\alias{convert.expression}
\title{converts expression data from a dataframe into an object.}
\usage{
convert.expression(data, entries = NULL)
}
\arguments{
\item{data}{A dataframe of expression data to be loaded.}

\item{entries}{An optional string array that designates
the entries of the dataframe to be loaded.}
}
\value{
An object that represents the expression data of microarray/RNA-seq.
    Internally, the object holds a matrix of size
    \code{nsamples * ngenes}.
}
\description{
converts expression data from a dataframe into an object.
}
