% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIT.R
\name{convert.weather}
\alias{convert.weather}
\title{Converts weather data from a dataframe into an object.}
\usage{
convert.weather(data, entries = IO$weather.entries)
}
\arguments{
\item{data}{A dataframe of weather data to be converted.}

\item{entries}{An optional string array that designates
the entries of the dataframe to be loaded.}
}
\value{
An object that reprents the timeseries data of weather factors.
    Internally, the object holds a dataframe of size
    \code{ntimepoints * nfactors}.
}
\description{
Converts weather data from a dataframe into an object.
}
