% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIT.R
\name{convert.weight}
\alias{convert.weight}
\title{Converts regression weight data from a dataframe into an object.}
\usage{
convert.weight(data, entries = NULL)
}
\arguments{
\item{data}{A dataframe that contains weight data to be loaded.}

\item{entries}{An optional string array that designates
the entries of the dataframe to be loaded.}
}
\value{
An object that represents the weights 
    Internally, the object holds a matrix of size
    \code{nsamples * ngenes}.
}
\description{
Converts regression weight data from a dataframe into an object.
}
