% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIT.R
\name{load.attribute}
\alias{load.attribute}
\title{Loads attribute data.}
\usage{
load.attribute(path, variable = NULL, sample = NULL)
}
\arguments{
\item{path}{A path of a file that contains attribute data to be loaded.
When the file is a loadable \code{.Rdata},
\code{name} of the dataframe object in the \code{.Rdata}
(that actually contains the relevant data)
has to be specified as well.}

\item{variable}{An optional string that designates the name of a
dataframe object that has been saved in an \code{.Rdata}.
(See the description of \code{path}.)}

\item{sample}{An optional numeric array that designates
the samples, that is rows, of the dataframe to be loaded.}
}
\value{
An object that represents the attributes of
    microarray/RNA-seq data.
    Internally, the object holds a dataframe whose number of entries
    (rows) equals that of the samples.
}
\description{
Loads attribute data.
}
