% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIT.R
\name{load.weight}
\alias{load.weight}
\title{Loads regression weight data.}
\usage{
load.weight(path, variable = NULL, entries = NULL)
}
\arguments{
\item{path}{A path of a file that contains weight data to be loaded.
When the file is a loadable \code{.Rdata},
\code{name} of the dataframe object in the \code{.Rdata}
(that actually contains the relevant data)
has to be specified as well.}

\item{variable}{An optional string that designates the name of a
dataframe object that has been saved in an \code{.Rdata}.
(See the description of \code{path}.)}

\item{entries}{An optional string array that designates
the entries of the dataframe to be loaded.}
}
\value{
An object that represents the weights 
    Internally, the object holds a matrix of size
    \code{nsamples * ngenes}.
}
\description{
Loads regression weight data.
}
