% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIT.R
\name{make.trivial.weights}
\alias{make.trivial.weights}
\title{Makes trivial weight data}
\usage{
make.trivial.weights(samples.n, genes)
}
\arguments{
\item{samples.n}{A number of samples.}

\item{genes}{A list of genes.}
}
\value{
An object that represens the trivial weights. 
       Internally, the object holds an identity matrix of size 
       \code{nsamples * ngenes}.
}
\description{
Makes trivial weight data
}
