% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIT.R
\docType{data}
\name{weather.entries}
\alias{weather.entries}
\title{Supported weather factors.}
\format{An object of class \code{character} of length 6.}
\usage{
weather.entries
}
\description{
Supported weather factors.
}
\examples{
\dontrun{
> length(FIT::weather.entries)
[1] 6
> FIT::weather.entries
[1] "wind"          "temperature"   "humidity"      "atmosphere"
[5] "precipitation" "radiation"
}
}
\keyword{datasets}

